// src/app/model/vendorFine.ts
export class VendorFine {

    id: number;
    dealerId: number;
    dealerName: string;
    status: string;
    poNumber: string;
    quarter: string;
    reason: string;
    contractNumber: string;
    checkNumber: string;
    checkDate: Date;
    comments: string;
    fineAmount: number;
    balance: number;
    fineAmountPaid: number;
    paymentDate: Date;
    updatedBy: string;
    updateDate: Date;
    createdBy: string;
    createdDate: Date;
    startDate: Date;

}